requirements.txt
aniso8601==8.0.0
apturl==0.5.2
asgiref==3.5.0
async-generator==1.10
attrs==19.3.0
Automat==0.8.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dlib==19.23.99
duplicity==0.8.12.0
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fasteners==0.14.1
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
future==0.18.2
h11==0.13.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
incremental==16.10.1
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
linecache2==1.0.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.1
monotonic==1.5
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.5.62
outcome==1.1.0
paramiko==2.6.0
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.8.0
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyperclip==1.8.2
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2020.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.27.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selenium==4.1.0
service-identity==18.1.0
simplejson==3.16.0
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.3.18
sqlparse==0.4.2
systemd-python==234
testresources==2.0.0
toml==0.10.2
traceback2==1.4.0
trio==0.19.0
trio-websocket==0.9.2
Twisted==18.9.0
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
Werkzeug==1.0.1
wsproto==1.0.0
xkit==0.0.0
youtube-dl==2020.3.24
zope.interface==4.7.1
